requirements.txt
openai
GitPython
flask
unittest
